+++ monitoring/rpc_stats_interceptor.go
@@ -24,8 +24,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-const traceSpanRoot = "/trillian/mon/"
-
 // RPCStatsInterceptor provides a gRPC interceptor that records statistics about the RPCs passing through it.
 type RPCStatsInterceptor struct {
 	prefix            string
@@ -70,11 +68,6 @@ func (r *RPCStatsInterceptor) recordFailureLatency(labels []string, startTime ti
 func (r *RPCStatsInterceptor) Interceptor() grpc.UnaryServerInterceptor {
 	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
 		labels := []string{info.FullMethod}
-
-		// This interceptor wraps the request handler so we should track the
-		// additional latency it imposes.
-		ctx, spanEnd := StartSpan(ctx, traceSpanRoot)
-		defer spanEnd()
 
 		// Increase the request count for the method and start the clock
 		r.ReqCount.Inc(labels...)